Documentation; Code Formatting (#108)
* cleanup before prettier

* format all contracts

* document interfaces

* add comments to major contracts; protocol fee testing

* prettier contracts

* exclude coverage

+++ contracts/libraries/utils/String.sol
@@ -2,17 +2,17 @@
 pragma solidity 0.8.18;
 
 library String {
-    bytes16 private constant alphabet = "0123456789abcdef";
+    bytes16 private constant alphabet = '0123456789abcdef';
 
-    function from(bytes32 value) internal pure returns(string memory) {
+    function from(bytes32 value) internal pure returns (string memory) {
         return toString(abi.encodePacked(value));
     }
 
-    function from(address account) internal pure returns(string memory) {
+    function from(address account) internal pure returns (string memory) {
         return toString(abi.encodePacked(account));
     }
 
-    function from(uint256 value) internal pure returns(string memory) {
+    function from(uint256 value) internal pure returns (string memory) {
         unchecked {
             uint256 length = log10(value) + 1;
             string memory buffer = new string(length);
@@ -35,7 +35,7 @@ library String {
     }
 
     function from(int256 value) internal pure returns (string memory) {
-        return string(abi.encodePacked(value < 0 ? "-" : "", from(abs(value))));
+        return string(abi.encodePacked(value < 0 ? '-' : '', from(abs(value))));
     }
 
     function abs(int256 n) internal pure returns (uint256) {
@@ -48,44 +48,44 @@ library String {
     function log10(uint256 value) internal pure returns (uint256) {
         uint256 result = 0;
         unchecked {
-            if (value >= 10 ** 64) {
-                value /= 10 ** 64;
+            if (value >= 10**64) {
+                value /= 10**64;
                 result += 64;
             }
-            if (value >= 10 ** 32) {
-                value /= 10 ** 32;
+            if (value >= 10**32) {
+                value /= 10**32;
                 result += 32;
             }
-            if (value >= 10 ** 16) {
-                value /= 10 ** 16;
+            if (value >= 10**16) {
+                value /= 10**16;
                 result += 16;
             }
-            if (value >= 10 ** 8) {
-                value /= 10 ** 8;
+            if (value >= 10**8) {
+                value /= 10**8;
                 result += 8;
             }
-            if (value >= 10 ** 4) {
-                value /= 10 ** 4;
+            if (value >= 10**4) {
+                value /= 10**4;
                 result += 4;
             }
-            if (value >= 10 ** 2) {
-                value /= 10 ** 2;
+            if (value >= 10**2) {
+                value /= 10**2;
                 result += 2;
             }
-            if (value >= 10 ** 1) {
+            if (value >= 10**1) {
                 result += 1;
             }
         }
         return result;
     }
 
-    function toString(bytes memory data) internal pure returns(string memory) {
+    function toString(bytes memory data) internal pure returns (string memory) {
         bytes memory str = new bytes(2 + data.length * 2);
-        str[0] = "0";
-        str[1] = "x";
-        for (uint i = 0; i < data.length;) {
-            str[2+i*2] = alphabet[uint(uint8(data[i] >> 4))];
-            str[3+i*2] = alphabet[uint(uint8(data[i] & 0x0f))];
+        str[0] = '0';
+        str[1] = 'x';
+        for (uint256 i = 0; i < data.length; ) {
+            str[2 + i * 2] = alphabet[uint256(uint8(data[i] >> 4))];
+            str[3 + i * 2] = alphabet[uint256(uint8(data[i] & 0x0f))];
             unchecked {
                 ++i;
             }
