+++ contracts/utils/LimitPoolManager.sol
@@ -9,7 +9,10 @@ import '../base/events/LimitPoolManagerEvents.sol';
 import '../libraries/utils/SafeCast.sol';
 
 /**
- * @dev Defines the actions which can be executed by the factory admin.
+ * @title LimitPoolManager
+ * @notice The manager for all limit pools
+ * @author Poolshark
+ * @author @alphak3y
  */
 contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
     address public owner;
@@ -18,8 +21,8 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
     // fee delta const for dynamic fees
     uint16 public feeDeltaConst;
     // max protocol fees
-    uint16  public constant MAX_PROTOCOL_SWAP_FEE = 1e4; /// @dev - max protocol swap fee of 100%
-    uint16  public constant MAX_PROTOCOL_FILL_FEE = 1e2; /// @dev - max protocol fill fee of 1%
+    uint16 public constant MAX_PROTOCOL_SWAP_FEE = 1e4; /// @dev - max protocol swap fee of 100%
+    uint16 public constant MAX_PROTOCOL_FILL_FEE = 1e2; /// @dev - max protocol fill fee of 1%
     // impl name => impl address
     bytes32[] _poolTypeNames;
     mapping(uint256 => address) internal _poolImpls;
@@ -75,12 +78,14 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
      * Can only be called by the current owner.
      */
     function transferOwner(address newOwner) public virtual onlyOwner {
-        if(newOwner == address(0)) require (false, 'TransferredToZeroAddress()');
+        if (newOwner == address(0))
+            require(false, 'TransferredToZeroAddress()');
         _transferOwner(newOwner);
     }
 
     function transferFeeTo(address newFeeTo) public virtual onlyOwner {
-        if(newFeeTo == address(0)) require (false, 'TransferredToZeroAddress()');
+        if (newFeeTo == address(0))
+            require(false, 'TransferredToZeroAddress()');
         _transferFeeTo(newFeeTo);
     }
 
@@ -104,10 +109,10 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
         emit FeeToTransfer(oldFeeTo, newFeeTo);
     }
 
-    function enableFeeTier(
-        uint16 swapFee,
-        int16 tickSpacing
-    ) external onlyOwner {
+    function enableFeeTier(uint16 swapFee, int16 tickSpacing)
+        external
+        onlyOwner
+    {
         if (_feeTiers[swapFee] != 0) revert FeeTierAlreadyEnabled();
         if (tickSpacing <= 0) revert InvalidTickSpacing();
         if (tickSpacing % 2 != 0) revert InvalidTickSpacing();
@@ -134,40 +139,40 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
         emit PoolTypeEnabled(poolTypeName_, poolImpl_, tokenImpl_, poolTypeId_);
     }
 
-    function setFactory(
-        address factory_
-    ) external onlyOwner {
-        if (factory != address(0)) require (false, 'FactoryAlreadySet()');
+    function setFactory(address factory_) external onlyOwner {
+        if (factory != address(0)) require(false, 'FactoryAlreadySet()');
         emit FactoryChanged(factory, factory_);
         factory = factory_;
     }
 
-    function setFeeDeltaConst(
-        uint16 feeDeltaConst_
-    ) external onlyOwner {
-        if (feeDeltaConst_ > 10000) require (false, 'FeeDeltaConstCeilingExceeded()');
+    function setFeeDeltaConst(uint16 feeDeltaConst_) external onlyOwner {
+        if (feeDeltaConst_ > 10000)
+            require(false, 'FeeDeltaConstCeilingExceeded()');
         emit FeeDeltaConstChanged(feeDeltaConst, feeDeltaConst_);
         feeDeltaConst = feeDeltaConst_;
     }
 
-    function collectProtocolFees(
-        address[] calldata pools
-    ) external onlyOwnerOrFeeTo {
-        if (pools.length == 0) require (false, 'EmptyPoolsArray()');
+    function collectProtocolFees(address[] calldata pools)
+        external
+        onlyOwnerOrFeeTo
+    {
+        if (pools.length == 0) require(false, 'EmptyPoolsArray()');
         uint128[] memory token0FeesCollected = new uint128[](pools.length);
         uint128[] memory token1FeesCollected = new uint128[](pools.length);
         // pass empty fees params
         FeesParams memory feesParams;
-        for (uint i; i < pools.length;) {
-            (
-                token0FeesCollected[i],
-                token1FeesCollected[i]
-            ) = IPool(pools[i]).fees(feesParams);
+        for (uint256 i; i < pools.length; ) {
+            (token0FeesCollected[i], token1FeesCollected[i]) = IPool(pools[i])
+                .fees(feesParams);
             unchecked {
                 ++i;
             }
         }
-        emit ProtocolFeesCollected(pools, token0FeesCollected, token1FeesCollected);
+        emit ProtocolFeesCollected(
+            pools,
+            token0FeesCollected,
+            token1FeesCollected
+        );
     }
 
     // protocol fee flags
@@ -180,9 +185,9 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
         address[] calldata pools,
         FeesParams[] calldata feesParams
     ) external onlyOwner {
-        if (pools.length == 0) require (false, 'EmptyPoolsArray()');
+        if (pools.length == 0) require(false, 'EmptyPoolsArray()');
         if (pools.length != feesParams.length) {
-            require (false, 'MismatchedArrayLengths()');
+            require(false, 'MismatchedArrayLengths()');
         }
         uint128[] memory token0FeesCollected = new uint128[](pools.length);
         uint128[] memory token1FeesCollected = new uint128[](pools.length);
@@ -190,13 +195,9 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
         int16[] memory protocolSwapFees1 = new int16[](pools.length);
         int16[] memory protocolFillFees0 = new int16[](pools.length);
         int16[] memory protocolFillFees1 = new int16[](pools.length);
-        for (uint i; i < pools.length;) {
-            (
-                token0FeesCollected[i],
-                token1FeesCollected[i]
-            ) = IPool(pools[i]).fees(
-                feesParams[i]
-            );
+        for (uint256 i; i < pools.length; ) {
+            (token0FeesCollected[i], token1FeesCollected[i]) = IPool(pools[i])
+                .fees(feesParams[i]);
             if ((feesParams[i].setFeesFlags & PROTOCOL_SWAP_FEE_0) > 0) {
                 protocolSwapFees0[i] = int16(feesParams[i].protocolSwapFee0);
             } else {
@@ -243,29 +244,27 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
         );
     }
 
-    function poolTypes(
-        uint16 poolTypeId
-    ) external view returns (
-        address,
-        address
-    ) {
+    function poolTypes(uint16 poolTypeId)
+        external
+        view
+        returns (address, address)
+    {
         return (_poolImpls[poolTypeId], _tokenImpls[poolTypeId]);
     }
 
-    function feeTiers(
-        uint16 swapFee
-    ) external view returns (
-        int16 tickSpacing
-    ) {
+    function feeTiers(uint16 swapFee)
+        external
+        view
+        returns (int16 tickSpacing)
+    {
         return _feeTiers[swapFee];
     }
-    
+
     /**
      * @dev Throws if the sender is not the owner.
      */
     function _checkOwner() internal view {
-        if (owner != msg.sender)
-            require (false, 'OwnerOnly()');
+        if (owner != msg.sender) require(false, 'OwnerOnly()');
     }
 
     /**
@@ -273,6 +272,6 @@ contract LimitPoolManager is ILimitPoolManager, LimitPoolManagerEvents {
      */
     function _checkFeeToAndOwner() internal view {
         if (feeTo != msg.sender && owner != msg.sender)
-            require (false, 'OwnerOrFeeToOnly()');
+            require(false, 'OwnerOrFeeToOnly()');
     }
 }
